Use C-Rate Providers

+++ src/pyprotolinc/portfolio.py
@@ -52,10 +52,10 @@ class Portfolio:
         # extract and validate the sex, map to enum coding
         self.gender = df_portfolio["SEX"].str.upper()
         assert set(self.gender.unique()).issubset({g.name for g in Gender})
-        self.gender = self.gender.map({g.name: int(g) for g in Gender}).values
+        self.gender = self.gender.map({g.name: int(g) for g in Gender}).values.astype(np.int)
 
         # extract age vector at the portfolio date
-        self.initial_ages = completed_months_to_date(df_portfolio["DATE_OF_BIRTH"], self.portfolio_date)
+        self.initial_ages = completed_months_to_date(df_portfolio["DATE_OF_BIRTH"], self.portfolio_date).astype(np.int)
         self.years_of_birth = self.df_portfolio["DATE_OF_BIRTH"].dt.year.values.astype(np.int16)
         self.months_of_birth = self.df_portfolio["DATE_OF_BIRTH"].dt.month.values.astype(np.int16)
 
@@ -78,10 +78,10 @@ class Portfolio:
         _unknown_smoker_status = set(self.smokerstatus) - {s.name for s in SmokerStatus}
         assert len(_unknown_smoker_status) == 0,\
             "Unknow smoker status: " + str(_unknown_smoker_status) + ", use one of " + str({s.name for s in SmokerStatus})
-        self.smokerstatus = self.smokerstatus.map(SmokerStatus.index_mapper())
+        self.smokerstatus = self.smokerstatus.map(SmokerStatus.index_mapper()).astype(np.int)
 
         # the number of month since the disablement date, is NaN if no disablement date is given
-        self.months_disabled_at_start = completed_months_to_date(df_portfolio["DATE_OF_DISABLEMENT"], self.portfolio_date)
+        self.months_disabled_at_start = completed_months_to_date(df_portfolio["DATE_OF_DISABLEMENT"], self.portfolio_date).astype(np.int)
 
         # check that when in disabled state at start then the disablement date must be at or before the portfolio_date
         disabled_according_to_date = df_portfolio.CURRENT_STATUS.str[:3] == "DIS"
